refactor: move configs into separate module
