+++ uplink/src/lib.rs
@@ -10,169 +10,15 @@ use tracing::error;
 
 pub mod base;
 pub mod collector;
+pub mod config;
 
-pub mod config {
-    use crate::base::StreamConfig;
-    pub use crate::base::{Config, Persistence, Stats};
-    use config::{Environment, File, FileFormat};
-    use std::fs;
-    use structopt::StructOpt;
-
-    #[derive(StructOpt, Debug)]
-    #[structopt(name = "uplink", about = "collect, batch, compress, publish")]
-    pub struct CommandLine {
-        /// Binary version
-        #[structopt(skip = env ! ("VERGEN_BUILD_SEMVER"))]
-        pub version: String,
-        /// Build profile
-        #[structopt(skip = env ! ("VERGEN_CARGO_PROFILE"))]
-        pub profile: String,
-        /// Commit SHA
-        #[structopt(skip = env ! ("VERGEN_GIT_SHA"))]
-        pub commit_sha: String,
-        /// Commit SHA
-        #[structopt(skip = env ! ("VERGEN_GIT_COMMIT_TIMESTAMP"))]
-        pub commit_date: String,
-        /// config file
-        #[structopt(short = "c", help = "Config file")]
-        pub config: Option<String>,
-        /// config file
-        #[structopt(short = "a", help = "Authentication file")]
-        pub auth: String,
-        /// log level (v: info, vv: debug, vvv: trace)
-        #[structopt(short = "v", long = "verbose", parse(from_occurrences))]
-        pub verbose: u8,
-        /// list of modules to log
-        #[structopt(short = "m", long = "modules")]
-        pub modules: Vec<String>,
-    }
-
-    const DEFAULT_CONFIG: &str = r#"
-    bridge_port = 5555
-    run_logcat = true
-    max_packet_size = 102400
-    max_inflight = 100
-    keep_alive = 60
-
-    # Whitelist of binaries which uplink can spawn as a process
-    # This makes sure that user is protected against random actions
-    # triggered from cloud.
-    actions = ["tunshell"]
-
-    # Create empty streams map
-    [streams]
-
-    [serializer_metrics]
-    enabled = false
-
-    [stream_metrics]
-    enabled = false
-
-    [action_status]
-    topic = "/tenants/{tenant_id}/devices/{device_id}/action/status"
-    buf_size = 1
-
-    [stats]
-    enabled = false
-    process_names = ["uplink"]
-    update_period = 30
-"#;
-
-    /// Reads config file to generate config struct and replaces places holders
-    /// like bike id and data version
-    pub fn initialize(auth_config: &str, uplink_config: &str) -> Result<Config, anyhow::Error> {
-        let config = config::Config::builder()
-            .add_source(File::from_str(DEFAULT_CONFIG, FileFormat::Toml))
-            .add_source(File::from_str(uplink_config, FileFormat::Toml))
-            .add_source(File::from_str(auth_config, FileFormat::Json))
-            .add_source(Environment::default())
-            .build()?;
-
-        let mut config: Config = config.try_deserialize()?;
-
-        if config.simulator.is_some() {
-            config.device_id = "+".to_string();
-        }
-
-        if let Some(persistence) = &config.persistence {
-            fs::create_dir_all(&persistence.path)?;
-        }
-
-        // replace placeholders with device/tenant ID
-        let tenant_id = config.project_id.trim();
-        let device_id = config.device_id.trim();
-        for config in config.streams.values_mut() {
-            replace_topic_placeholders(config, tenant_id, device_id);
-        }
-
-        replace_topic_placeholders(&mut config.action_status, tenant_id, device_id);
-
-        for config in [&mut config.serializer_metrics, &mut config.stream_metrics] {
-            if let Some(topic) = &config.topic {
-                let topic = topic.replace("{tenant_id}", tenant_id);
-                let topic = topic.replace("{device_id}", device_id);
-                config.topic = Some(topic);
-            }
-        }
-
-        for stat in [
-            "disk_stats",
-            "network_stats",
-            "processor_stats",
-            "process_stats",
-            "component_stats",
-            "system_stats",
-        ] {
-            config.bypass_streams.push("uplink_".to_string() + stat);
-        }
-
-        Ok(config)
-    }
-
-    // Replace placeholders in topic strings with configured values for tenant_id and device_id
-    fn replace_topic_placeholders(config: &mut StreamConfig, tenant_id: &str, device_id: &str) {
-        if let Some(topic) = &config.topic {
-            let topic = topic.replace("{tenant_id}", tenant_id);
-            let topic = topic.replace("{device_id}", device_id);
-            config.topic = Some(topic);
-        }
-    }
-
-    #[derive(Debug, thiserror::Error)]
-    pub enum ReadFileError {
-        #[error("Auth file not found at {0}")]
-        Auth(String),
-        #[error("Config file not found at {0}")]
-        Config(String),
-    }
-
-    fn read_file_contents(path: &str) -> Option<String> {
-        fs::read_to_string(path).ok()
-    }
-
-    pub fn get_configs(
-        commandline: &CommandLine,
-    ) -> Result<(String, Option<String>), ReadFileError> {
-        let auth = read_file_contents(&commandline.auth)
-            .ok_or_else(|| ReadFileError::Auth(commandline.auth.to_string()))?;
-        let config = match &commandline.config {
-            Some(path) => Some(
-                read_file_contents(path).ok_or_else(|| ReadFileError::Config(path.to_string()))?,
-            ),
-            None => None,
-        };
-
-        Ok((auth, config))
-    }
-}
-
+use crate::config::Config;
 pub use base::actions::{Action, ActionResponse};
-pub use base::middleware;
 use base::middleware::tunshell::TunshellSession;
 use base::middleware::Middleware;
 use base::mqtt::Mqtt;
 use base::serializer::Serializer;
-pub use base::{Config, Package, Payload, Point, Stream};
+pub use base::{middleware, Package, Payload, Point, Stream};
 use collector::downloader::FileDownloader;
 use collector::systemstats::StatCollector;
 pub use collector::{simulator, tcpjson::Bridge};
